refactor(ui): share task progress across GUI and WebUI
