+++ crates/agent-gateway/web/test/task-progress-indicator.test.mjs
@@ -7,6 +7,12 @@ import { createWebModuleLoader } from "../../test/helpers/load-web-module.mjs";
 const rootDir = fileURLToPath(new URL("../", import.meta.url));
 const iconsPath = fileURLToPath(new URL("../src/components/icons/index.ts", import.meta.url));
 const utilsPath = fileURLToPath(new URL("../src/lib/shared/utils.ts", import.meta.url));
+const localeContextPath = fileURLToPath(
+  new URL("../../../agent-ui/src/i18n/LocaleContext.tsx", import.meta.url),
+);
+const taskProgressIndicatorPath = fileURLToPath(
+  new URL("../../../agent-ui/src/components/chat/TaskProgressIndicator.tsx", import.meta.url),
+);
 
 const labels = {
   title: "Task progress",
@@ -377,4 +383,35 @@ test("web shows pending, paused, and completed states without auto-dismissing co
   });
   assert.match(treeText(indicator.render({ snapshot: completed })), /All completed/);
   assert.match(treeText(indicator.render({ snapshot: completed })), /All completed/);
+});
+
+test("web uses the shared localized task progress bar", () => {
+  const indicator = (props) => ({ type: "TaskProgressIndicator", props });
+  const translations = {
+    "chat.taskProgress.title": "Task progress",
+    "chat.taskProgress.step": "Step {current} of {total}",
+    "chat.taskProgress.completedCount": "completed",
+    "chat.taskProgress.running": "Running",
+    "chat.taskProgress.pending": "Pending",
+    "chat.taskProgress.paused": "Paused",
+    "chat.taskProgress.completed": "All completed",
+  };
+  const loader = createWebModuleLoader({
+    rootDir,
+    mocks: {
+      [localeContextPath]: {
+        useLocale: () => ({ t: (key) => translations[key] ?? key }),
+      },
+      [taskProgressIndicatorPath]: { TaskProgressIndicator: indicator },
+    },
+  });
+  const { TaskProgressBar } = loader.loadModule(
+    "@liveagent/ui/components/chat/TaskProgressBar.tsx",
+  );
+  const snapshot = createSnapshot();
+  const tree = TaskProgressBar({ snapshot, isConversationRunning: true });
+
+  assert.equal(tree.type, indicator);
+  assert.deepEqual(tree.props.labels, labels);
+  assert.equal(TaskProgressBar({ snapshot: null, isConversationRunning: false }), null);
 });
